# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /geofences:
    get:
      operationId: getGeofences
      tags: [geofences]
      summary: List geofences
      responses:
        "200":
          description: Geofences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Geofence"
    post:
      operationId: createGeofence
      tags: [geofences]
      summary: Create a geofence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeofenceInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Geofence"
  /geofences/{id}:
    patch:
      operationId: updateGeofence
      tags: [geofences]
      summary: Update a geofence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeofenceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Geofence"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
    delete:
      operationId: deleteGeofence
      tags: [geofences]
      summary: Delete a geofence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /dispatch/jobs:
    get:
      operationId: getDispatchJobs
      tags: [dispatch]
      summary: List dispatch jobs
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DispatchJobItem"
    post:
      operationId: createDispatchJob
      tags: [dispatch]
      summary: Create a dispatch job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchJobInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchJobItem"
  /dispatch/jobs/{id}:
    patch:
      operationId: updateDispatchJob
      tags: [dispatch]
      summary: Update a dispatch job (assign, progress status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchJobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchJobItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
  /dispatch/jobs/{id}/messages:
    get:
      operationId: getDispatchMessages
      tags: [dispatch]
      summary: List messages for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DispatchMessageItem"
    post:
      operationId: sendDispatchMessage
      tags: [dispatch]
      summary: Send a message to the driver on a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchMessageInput"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchMessageItem"
...
components:
  schemas:
...
    ApiErrorMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Geofence:
      type: object
      required: [id, name, kind, centerLat, centerLng, radiusM, active]
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
        centerLat:
          type: number
        centerLng:
          type: number
        radiusM:
          type: integer
        active:
          type: boolean
    GeofenceInput:
      type: object
      required: [name, kind, centerLat, centerLng, radiusM]
      properties:
        name:
          type: string
          minLength: 1
        kind:
          type: string
          enum: [DEPOT, CORRIDOR, BORDER, CUSTOMER, NOGO]
        centerLat:
          type: number
        centerLng:
          type: number
        radiusM:
          type: integer
          minimum: 50
        active:
          type: boolean
    GeofenceUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        kind:
          type: string
          enum: [DEPOT, CORRIDOR, BORDER, CUSTOMER, NOGO]
        centerLat:
          type: number
        centerLng:
          type: number
        radiusM:
          type: integer
          minimum: 50
        active:
          type: boolean
...
    DispatchJobItem:
      type: object
      required: [id, reference, origin, destination, status, priority, createdAt, unreadMessages]
      properties:
        id:
          type: string
        reference:
          type: string
        driverId:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
        vehicleId:
          type: ["string", "null"]
        vehicleReg:
          type: ["string", "null"]
        origin:
          type: string
        destination:
          type: string
        status:
          type: string
        priority:
          type: string
        cargo:
          type: ["string", "null"]
        weightKg:
          type: ["integer", "null"]
        distanceKm:
          type: ["integer", "null"]
        scheduledFor:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        createdAt:
          type: string
        unreadMessages:
          type: integer
    DispatchJobInput:
      type: object
      required: [reference, origin, destination]
      properties:
        reference:
          type: string
          minLength: 1
        origin:
          type: string
          minLength: 1
        destination:
          type: string
          minLength: 1
        driverId:
          type: string
        vehicleId:
          type: string
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        cargo:
          type: string
        weightKg:
          type: integer
        distanceKm:
          type: integer
        scheduledFor:
          type: string
    DispatchJobUpdate:
      type: object
      properties:
        driverId:
          type: ["string", "null"]
        vehicleId:
          type: ["string", "null"]
        status:
          type: string
          enum: [DRAFT, ASSIGNED, EN_ROUTE, DELIVERED, CANCELLED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        cargo:
          type: string
        weightKg:
          type: integer
        distanceKm:
          type: integer
        scheduledFor:
          type: string
    DispatchMessageItem:
      type: object
      required: [id, direction, body, sentAt]
      properties:
        id:
          type: string
        direction:
          type: string
        body:
          type: string
        sentAt:
          type: string
        readAt:
          type: ["string", "null"]
    DispatchMessageInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
...