# lib/api-zod/src/generated/types/dispatchJobUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DispatchJobUpdatePriority } from './dispatchJobUpdatePriority';
import type { DispatchJobUpdateStatus } from './dispatchJobUpdateStatus';

export interface DispatchJobUpdate {
  /** @nullable */
  driverId?: string | null;
  /** @nullable */
  vehicleId?: string | null;
  status?: DispatchJobUpdateStatus;
  priority?: DispatchJobUpdatePriority;
  cargo?: string;
  weightKg?: number;
  distanceKm?: number;
  scheduledFor?: string;
}